Change crowdinverify to use member and enable remote verification
May still need more testing with the error checking when remotely verifying

+++ commands/Utility/verify.js
@@ -6,11 +6,11 @@ const Discord = require("discord.js")
 module.exports = {
     name: "verify",
     description: "Unverifies the user.",
-    usage: "+verify",
+    usage: "+verify [profileURL]",
     aliases: ["unverify", "reverify"],
     cooldown: 3600,
     allowTip: false,
-    async execute(message, args, getString) {
+    async execute(message, args) {
         if (!message.member.roles.cache.has("569194996964786178")) { //Verified
             await message.delete()
             message.channel.messages.fetch()
@@ -21,12 +21,13 @@ module.exports = {
             await message.member.roles.add("569194996964786178", "Manually verified through the command")
                 .then(async () => await message.member.roles.remove("756199836470214848", "Manually verified through the command")) //Add Verified and remove Alerted
             message.guild.channels.cache.get("662660931838410754").send(`${message.author} manually verified themselves through the command`) //verify-logs
-        } else {
+        } else if (!message.member.roles.cache.has("764442984119795732") || !args[0]) { //Discord Administrator
             const userDb = await getDb().collection("users").findOne({ id: message.author.id })
             if (userDb.profile) {
                 message.react("798339571531382874") //icon_working
                 message.client.channels.cache.get("662660931838410754").send(`${message.author} was unverified.`) //verify-logs
-                return crowdinVerify(message)
+                await crowdinVerify(message.member, message.content.match(/(https:\/\/)([a-z]{2,}\.)?crowdin\.com\/profile\/\S{1,}/gi)?.[0], true)
+                message.delete()
             } else {
                 await message.member.roles.remove("569194996964786178", "Unverified") //verified
                 if (!message.deleted) message.delete()
@@ -51,6 +52,14 @@ module.exports = {
                         message.client.channels.cache.get("662660931838410754").send(`${message.author} was unverified and had DMs off.`) //verify-logs
                     })
             }
+        } else {
+            if (!args[0]) throw "noUser"
+            const user = args[0].replace(/[\\<>@&!]/g, "").toLowerCase()
+            const member = message.guild.members.cache.find(m => m.id === user || m.user.username.toLowerCase() === user || m.user.tag.toLowerCase() === user)
+            if (!member) throw "falseUser"
+            message.react("798339571531382874") //icon_working
+            await crowdinVerify(message.member, message.content.match(/(https:\/\/)([a-z]{2,}\.)?crowdin\.com\/profile\/\S{1,}/gi)?.[0])
+            message.delete()
         }
     }
 }
